+++ src/app/app.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { HeaderComponent } from "@core/layout/header/header.component";
-import { FooterComponent } from '@core/layout/footer/footer.component';
+import { HeaderComponent } from "@core/components/header/header.component";
+import { FooterComponent } from '@core/components/footer/footer.component';
 import { CommonModule } from '@angular/common';
 import { ToastComponent } from "./shared/components/toast/toast.component";
