+++ src/app/core/services/error-handler.service.ts
@@ -1,6 +1,6 @@
 import { HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { SweetAlertService } from '@shared/services/sweetAlert.service';
+import { SweetAlertService } from '@shared/services/sweet-alert.service';
 import { SweetAlertResult } from 'sweetalert2';
 
 @Injectable({ providedIn: 'root' })
