--- src/app/features/administrador/models/IAdmin.model.ts
@@ -1,5 +1,0 @@
-export interface IAdmin{
-    id : number,
-    userId : number,
-    nivel : number,
-}
