+++ src/app/features/reclamacao/pages/reclamacao-form/reclamacao-form.component.ts
@@ -1,4 +1,4 @@
-import { SweetAlertService } from '@shared/services/sweetAlert.service';
+import { SweetAlertService } from '@shared/services/sweet-alert.service';
 import { CommonModule } from '@angular/common';
 import { Component, inject, OnInit } from '@angular/core';
 import {
