+++ src/app/shared/enums/registration-error.enum.ts
@@ -3,7 +3,7 @@
    *invalidControl -> Campo inválido
    *nvalidPassword -> Senhas não coincidem
  */
-export enum CadastroErrorStatus {
+export enum RegistrationErrorEnum {
   None = "none",
   invalidPassword = "invalid-password",
   invalidControl = "invalid-control",
