+++ src/app/shared/models/search-feedback.model.ts
