+++ src/app/shared/services/sweet-alert.service.ts
